Typescript bug fixes

+++ src/components/MyDoings.tsx
@@ -5,34 +5,28 @@ import Button from "@mui/material/Button";
 import CircularProgress from "@mui/material/CircularProgress";
 
 const MyDoings: FC = () => {
-  const focusRef = useRef();
-  interface doings {
-    title: string;
-    newDoings: string | number;
-    date: Date;
-  }
   const {
     todos,
     handleDelete,
-    inputRef,
     languages,
     translatedText,
     textToTranslate,
     setChangeLanguage,
     setTextToTranslate,
     isFetching,
-    sourceLang,
     setSourceLang,
-  } = useContext(myContextApi);
-  const [sureToDelete, setSureToDelete] = useState<boolean>(false);
+  } = useContext(myContextApi) ?? {};
 
-  console.log(translatedText);
-  console.log(textToTranslate);
-  //
+  const [sureToDelete, setSureToDelete] = useState<boolean>(false);
+  const handleLanguageChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
+    const selectedLang = e.target.value;
+    setChangeLanguage && setChangeLanguage(selectedLang);
+  };
 
   return (
     <div className="flex flex-col gap-5 w-[90%] md:w-[1/2]">
-      {todos.map((doings: doings, index: number) => {
+      {todos?.map((doings, index: number) => {
+        console.log(doings.date ? doings.date : null);
         return (
           <div
             className="relative m-auto w-full shadow-2xl p-4 rounded-xl"
@@ -42,16 +36,22 @@ const MyDoings: FC = () => {
               <form className="flex flex-col items-center relative leading-tight justify-between gap-2">
                 <aside className="flex flex-col justify-between w-full gap-4">
                   <div className="flex justify-between items-center">
-                    <i className="text-[12px]">{doings.date}</i>
+                    <i className="text-[12px]">
+                      {doings.date
+                        ? doings.date.toLocaleString().slice(0, 17).split("T") +
+                          " " +
+                          " "
+                        : ""}
+                    </i>
                     <select
                       name="country"
                       required
                       className="border bg-black/50 text-gray-200 text-[12px] md:text-md rounded"
-                      onChange={(e) => setChangeLanguage(e.target.value)}
+                      onChange={handleLanguageChange}
                     >
                       <option value="en">Translate Doings</option>
-                      {languages.map((language: string, index: number) => {
-                        setSourceLang(language.code);
+                      {languages?.map((language, index: number) => {
+                       setSourceLang && setSourceLang(language.code)
                         return (
                           <option key={index} value={language.code}>
                             {language.name}
@@ -74,18 +74,17 @@ const MyDoings: FC = () => {
                       disabled
                     />
                     {textToTranslate === doings.newDoings && isFetching && (
-                    <figure className="absolute w-full border h-full flex items-center justify-center">
-                      <CircularProgress
-                        color="success"
-                        
-                      />
-                    </figure>)}
+                      <figure className="absolute w-full border h-full flex items-center justify-center">
+                        <CircularProgress color="success" />
+                      </figure>
+                    )}
                   </div>
                 </aside>
                 {sureToDelete && (
                   <div className="text-start text-gray-800">
                     <p className="">
-                      Are you sure you want to delete this doings?
+                      Are you sure you want to delete this{" "}
+                      <strong>doings</strong>?
                     </p>{" "}
                     <div className="flex justify-between">
                       <button
@@ -97,7 +96,7 @@ const MyDoings: FC = () => {
                       <button
                         className="border p-1 rounded mt-2"
                         onClick={() => {
-                          handleDelete(index);
+                          handleDelete && handleDelete(index);
                           setSureToDelete(false);
                         }}
                       >
@@ -126,20 +125,6 @@ const MyDoings: FC = () => {
                       Delete
                     </Button>
 
-                    {/* <Button
-                      variant="contained"
-                      sx={{
-                        width: "10%",
-                        backgroundColor: "#D8605B",
-                        marginTop: "16px",
-                        fontSize: "10px",
-                        fontWeight: "bold",
-                        marginRight: "auto",
-                      }}
-                      onClick={editDoings}
-                    >
-                      Update
-                    </Button> */}
                     <Button
                       variant="contained"
                       sx={{
@@ -152,7 +137,7 @@ const MyDoings: FC = () => {
                         fontWeight: "bold",
                       }}
                       onClick={() => {
-                        setTextToTranslate(doings.newDoings);
+                        setTextToTranslate && setTextToTranslate(doings.newDoings);
                       }}
                     >
                       Translate
